requirements.txt
fastapi
uvicorn
firecrawl-py
supabase
openai
pydantic
python-dotenv
pydantic-settings
